Remove auto-prefixing from channel initialization
Previously, the topic set in your markup did not require a prefix. For
example, the topic `data-topic="posts"` would subscribe to the topic
`data:posts` implicitly. This caused a bit of confusion, because as
things are currently implemented, you have to use the prefix when
broadcasting. This changes keeps things consistent until an abstraction
is established for dealing with topics.